+++ RELEASE_NOTES.md
@@ -10,6 +10,8 @@ A new header is inserted each time a *tag* is created.
 - Added multi-bounce ambient occlusion to brighten AO and preserve local color (see `multiBounceAmbientOcclusion` property in materials)
 - Micro-shadowing is now applied to material ambient occlusion
 - Use a smaller 64x64 DFG LUT on mobile to reduce binary size
+- Added a distance field generator to libimage.
+- JavaScript MaterialInstance now supports vec4 colors.
 - Further reduced `filamat` binary size by removing reliance on stdlib.
 - Added a new, smaller, version of the `filamat` library, `filamat_lite`. Material optimization and
   compiling for non-OpenGL backends have been removed in favor of a smaller binary size.